+++ src/CssHslColor/CssHslColor.ts
@@ -39,6 +39,9 @@ import type { CssCmykColor } from "../CssCmykColor/CssCmykColor";
 import type { CssCmykColorData } from "../CssCmykColor/CssCmykColorData.type";
 import { makeCssCmykColorFromCssColor } from "../CssCmykColor/makeCssCmykColorFromCssColor";
 import { CssColor } from "../CssColor/CssColor";
+import type { CssHsvColor } from "../CssHsvColor/CssHsvColor";
+import type { CssHsvColorData } from "../CssHsvColor/CssHsvColorData.type";
+import { makeCssHsvColorFromCssColor } from "../CssHsvColor/makeCssHsvColorFromCssColor";
 import { CssHwbColor } from "../CssHwbColor/CssHwbColor";
 import type { CssHwbColorData } from "../CssHwbColor/CssHwbColorData.type";
 import { makeCssHwbColorFromCssColor } from "../CssHwbColor/makeCssHwbColorFromCssColor";
@@ -89,6 +92,21 @@ export class CssHslColor extends CssColor<CssHslColorData, Hsl>
     ): CssHslColor
     {
         return makeCssHslColorFromCssColor(
+            this,
+            { path, onError },
+            ...fnOpts
+        );
+    }
+
+    public hsv(
+        {
+            path = DEFAULT_DATA_PATH,
+            onError = THROW_THE_ERROR
+        }: DataGuaranteeOptions = {},
+        ...fnOpts: FunctionalOption<CssHsvColorData, DataGuaranteeOptions>[]
+    ): CssHsvColor
+    {
+        return makeCssHsvColorFromCssColor(
             this,
             { path, onError },
             ...fnOpts