+++ src/CssHwbColor/_fixtures/CssHwbColorData.ts
@@ -36,6 +36,7 @@ import type { Maybe } from "@safelytyped/core-types";
 import type { CssCmykColorChannelsData } from "../../CssCmykColor/CssCmykColorChannelsData.type";
 import { ValidCssColors } from "../../CssColor/_fixtures/CssColorFixtures";
 import type { CssHslColorChannelsData } from "../../CssHslColor/CssHslColorChannelsData.type";
+import type { CssHsvColorChannelsData } from "../../CssHsvColor/CssHsvColorChannelsData.type";
 import type { CssOklchColorChannelsData } from "../../CssOklchColor/CssOklchColorChannelsData.type";
 import type { CssRgbColorChannelsData } from "../../CssRgbColor/CssRgbColorChannelsData.type";
 import type { CssHwbColorData } from "../CssHwbColorData.type";
@@ -43,6 +44,7 @@ import type { CssHwbColorData } from "../CssHwbColorData.type";
 type ValidCssHwbColor = CssHwbColorData & {
     cmykChannels: CssCmykColorChannelsData;
     hslChannels: CssHslColorChannelsData;
+    hsvChannels: CssHsvColorChannelsData;
     rgbChannels: CssRgbColorChannelsData;
     oklchChannels: CssOklchColorChannelsData;
     hex: string;
@@ -61,6 +63,7 @@ ValidCssColors.forEach((fixture) => {
             colorSpace: "sRGB",
             cmykChannels: fixture.cmykChannels,
             hslChannels: fixture.hslChannels,
+            hsvChannels: fixture.hsvChannels,
             rgbChannels: fixture.rgbChannels,
             oklchChannels: fixture.oklchChannels,
             hex: fixture.hex,